Interim commit:  fill(..) test & impl
