+++ src/refill.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * This module provides a mixin for a Backbone.Collection to provide a method,
  * 'fill(...)' that can be used as a performant replacement for
@@ -6,92 +8,7 @@
 
 var _ = require('underscore');
 var Backbone = require('backbone');
-
-/**
- * This method is used as a replacement for the Backbone.Model constructor.  It allows
- * us to only calculate default values when requested.
- */
-var QuickModelConstructor = function(attributes, options) {
-    var attrs = attributes || {};
-    options || (options = {});
-    //noinspection JSUnusedGlobalSymbols
-    this.cid = _.uniqueId('c');
-    this.attributes = {};
-    if (options.collection) this.collection = options.collection;
-    if (options.parse) attrs = this.parse(attrs, options) || {};
-
-    // One significant change from Backbone.Model: only do defaults if necessary
-    var defaults = _.result(this, 'defaults');
-    if (defaults) {
-        attrs = _.defaults({}, attrs, _.result(this, 'defaults'));
-    }
-
-    this.set(attrs, options);
-    //noinspection JSUnusedGlobalSymbols
-    this.changed = {};
-    this.initialize.apply(this, arguments);
-};
-_.extend(QuickModelConstructor.prototype, Backbone.Model.prototype);
-
-/**
- * This function is swapped into a Backbone.Model's prototype when models are going to be
- * added to a collection in order to not do unnecessary work.
- */
-function quickModelSet(key, val) {
-    // Just assign the attribute & move on.
-    var attrs, current;
-    if (key == null) return this;
-
-    // Handle both `"key", value` and `{key: value}` -style arguments.
-    if (typeof key === 'object') {
-        attrs = key;
-    } else {
-        (attrs = {})[key] = val;
-    }
-
-    // Check for changes of `id`.
-    if (this.idAttribute in attrs) {
-        this.id = attrs[this.idAttribute];
-    }
-
-    // NOTE:  no validation, un-setting, _previousAttributes updating
-    current = this.attributes;
-    for (var attr in attrs) {
-        // NOTE:  no changes detection & event triggering
-
-        //noinspection JSUnfilteredForInLoop
-        val = attrs[attr];
-
-        //noinspection JSUnfilteredForInLoop
-        current[attr] = val;
-    }
-
-    return this;
-}
-
-/**
- * This function is used in place of the Backbone.Collection.set(...).
- * @param models
- * @param options
- * @returns {*}
- */
-function quickCollectionSet(models, options) {
-    // Force no-sort up front
-    var needsSort = options.sort;
-    if (options.sort) {
-        options.sort = false;
-    }
-
-    var returnedModels = this._originalCollectionSet(models, options);
-
-    // Handle sorting after we have set everything
-    if (needsSort && _.isArray(returnedModels)) {
-        this.sort();
-        returnedModels = _.clone(this.models);
-    }
-
-    return returnedModels;
-}
+var shortCircuit = require('./shortCircuit');
 
 function refill(models, options) {
 
@@ -99,19 +16,19 @@ function refill(models, options) {
     // original model Constructor
     var originalModelConstructor = this.model;
     if (_.isFunction(this.model.parse)) {
-        QuickModelConstructor.prototype.parse = this.model.prototype.parse;
+        shortCircuit.ModelConstructor.prototype.parse = this.model.prototype.parse;
     } else {
-        QuickModelConstructor.prototype.parse = Backbone.Model.prototype.parse;
+        shortCircuit.ModelConstructor.prototype.parse = Backbone.Model.prototype.parse;
     }
-    this.model = QuickModelConstructor;
+    this.model = shortCircuit.ModelConstructor;
 
     // Re-assign the Backbone.Model.set method
     var originalModelSet = this.model.prototype.set;
-    this.model.prototype.set = quickModelSet;
+    this.model.prototype.set = shortCircuit.modelSet;
 
     // Re-assign the Backbone.Collection.set method
     this._originalCollectionSet = this.set;
-    this.set = quickCollectionSet;
+    this.set = shortCircuit.collectionSet;
 
     // Call reset
     var result = this.reset(models, options);
@@ -134,10 +51,9 @@ var mixinObj = {
     refill: refill
 };
 
-
 module.exports = {
     mixin: function(Collection) {
-        _.extend(Collection.prototype, mixinObj );
+        _.extend(Collection.prototype, mixinObj);
         return Collection;
     }
 };
